NODE-911: Extract `CasperLabsProtocol` to separate file.
